Agrega fase de análisis semántico
Implementa una fase de análisis semántico para validar el árbol de sintaxis abstracta (AST).
Esto incluye el manejo de una tabla de símbolos para las entidades declaradas
(parámetros, fuentes, datasets, transforms, sinks) y la validación
de las dependencias de flujo de datos.

El analizador semántico verifica declaraciones duplicadas y asegura
que los datasets y transforms se utilicen con tipos de origen compatibles.

Además, se modificaron los permisos de ejecución de los scripts bash.

+++ src/main/c/support/type/TokenLabel.h
@@ -1,10 +1,15 @@
 #ifndef TOKEN_LABEL_HEADER
 #define TOKEN_LABEL_HEADER
 
+#include <stdint.h>
+
 /**
- * The type of a Bison token label, that is, an identifier of a token, that
- * transports the lexeme and semantic value as a whole.
+ * The type of a Bison token label. Use a pointer-sized signed integer so we
+ * can safely store pointers (casted) in token values on both 32- and
+ * 64-bit platforms. This project uses the token value to transport a
+ * lexeme pointer from the lexer to the parser for identifiers and
+ * string literals.
  */
-typedef signed int TokenLabel;
+typedef intptr_t TokenLabel;
 
 #endif
